% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{add_patterns}
\alias{add_patterns}
\title{Add Patterns in the table}
\usage{
add_patterns(Trend, Type, Open, High, Low, Close)
}
\arguments{
\item{Trend}{What is the trend currently}

\item{Type}{Type of pattern name}

\item{Open}{Open Price}

\item{High}{Close Price}

\item{Low}{Low Price}

\item{Close}{Close Price}
}
\value{
a table with pattern
}
\description{
Add Patterns in the table
}
