% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time-series-agg.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate df with the series to different period lengths}
\usage{
aggregate_data(df, period)
}
\arguments{
\item{df}{data frame to aggregate}

\item{period}{period day/month}
}
\value{
df
}
\description{
Aggregate df with the series to different period lengths
}
