% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{calculate_9_SMA_gt_20}
\alias{calculate_9_SMA_gt_20}
\title{Calculate is 9 SMA greater than 20 SMA}
\usage{
calculate_9_SMA_gt_20(df)
}
\arguments{
\item{df}{data frame}
}
\value{
return TRUE if 9 SMA is greater than 20 SMA else FALSE
}
\description{
Calculate is 9 SMA greater than 20 SMA
}
