% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{calculate_EMA}
\alias{calculate_EMA}
\title{Calculate Exponential Moving Average}
\usage{
calculate_EMA(df, ma_days = 9)
}
\arguments{
\item{df}{data frame}

\item{ma_days}{number of days}
}
\value{
Exponential Moving Average
}
\description{
Calculate Exponential Moving Average
}
