% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{calculate_SMA}
\alias{calculate_SMA}
\title{Calculate Simple Moving Average}
\usage{
calculate_SMA(df, ma_days = 9)
}
\arguments{
\item{df}{data frame}

\item{ma_days}{number of days}
}
\value{
Simple Moving Average
}
\description{
Calculate Simple Moving Average
}
