% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{calculate_technical_indicators}
\alias{calculate_technical_indicators}
\title{Calculate Technical Indicators}
\usage{
calculate_technical_indicators(df)
}
\arguments{
\item{df}{data frame}
}
\value{
return different technical indicator signals
}
\description{
Calculate Technical Indicators
}
