% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output_formating.R
\name{cell_color}
\alias{cell_color}
\title{Function to color the cell base on the cell value}
\usage{
cell_color(value, reverse = FALSE)
}
\arguments{
\item{value}{cell value}

\item{reverse}{to check the what color to apply}
}
\value{
return a color code
}
\description{
Function to color the cell base on the cell value
}
