% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time-series-agg.R
\name{day_num}
\alias{day_num}
\alias{year_num}
\title{Create grouping column based on day/wee/month/quarter/year}
\usage{
day_num(date_vec)

year_num(date_vec)
}
\arguments{
\item{date_vec}{Dates vector}
}
\description{
The function returns unique numbers of day/week/month/quarter taking into account the year.
The output can directly be used for creating aggregation groups.
}
