% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{detect_patterns}
\alias{detect_patterns}
\title{Function to detect specific patterns}
\usage{
detect_patterns(
  Trend,
  Type,
  Open1,
  Close1,
  High1,
  Low1,
  Open2,
  Close2,
  High2,
  Low2
)
}
\arguments{
\item{Trend}{Market trend}

\item{Type}{Type of candlestick pattern}

\item{Open1}{Open price of first candle}

\item{Close1}{Close price of first candle}

\item{High1}{High price of first candle}

\item{Low1}{Low price of first candle}

\item{Open2}{Open price of second candle}

\item{Close2}{Close price of second candle}

\item{High2}{High price of second candle}

\item{Low2}{Low price of second candle}
}
\value{
A data frame with patterns
}
\description{
Function to detect specific patterns
}
