% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_send_otp.R
\name{generate_otp}
\alias{generate_otp}
\title{Function to generate OTP}
\usage{
generate_otp(n = 6)
}
\arguments{
\item{n}{number of digits in OTP}
}
\value{
A OTP with 6 digits
}
\description{
Function to generate OTP
}
