% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_send_otp.R
\name{generate_password}
\alias{generate_password}
\title{Function to generate unique password}
\usage{
generate_password()
}
\value{
A unique password
}
\description{
Function to generate unique password
}
