% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_send_otp.R
\name{generate_user_name}
\alias{generate_user_name}
\title{Function to generate user name}
\usage{
generate_user_name(first_name, last_name)
}
\arguments{
\item{first_name}{Users first name}

\item{last_name}{Users last name}
}
\value{
A list of user name and password
}
\description{
Function to generate user name
}
