% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_buy_signals}
\alias{get_buy_signals}
\title{Calculate the Buy signals}
\usage{
get_buy_signals(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame with Buy Signals
}
\description{
Calculate the Buy signals
}
