% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_data}
\alias{get_data}
\title{Get the data stock from yahoo!}
\usage{
get_data(symbol, type, start_date, end_date)
}
\arguments{
\item{symbol}{symbol of the stock}

\item{type}{NSE or BSE}

\item{start_date}{Start Date}

\item{end_date}{End Date}
}
\value{
Returns a data of selected stock
}
\description{
Get the data stock from yahoo!
}
