% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{nadaraya_watson_envelope}
\alias{nadaraya_watson_envelope}
\title{Function to compute LuxAlgo Nadaraya-Watson Envelope and generate signals}
\usage{
nadaraya_watson_envelope(price, bandwidth = 8, alpha = 0.02)
}
\arguments{
\item{price}{close price of the stock}

\item{bandwidth}{set bandwidth}

\item{alpha}{set the alpha level}
}
\value{
returns a signals whether to buy or sell
}
\description{
Function to compute LuxAlgo Nadaraya-Watson Envelope and generate signals
}
