% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time-series-agg.R
\name{quarter_num}
\alias{quarter_num}
\title{Aggregate data on quarterly basis}
\usage{
quarter_num(date_vec)
}
\arguments{
\item{date_vec}{Dates vector}
}
\description{
Aggregate data on quarterly basis
}
