% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output_formating.R
\name{reactable_formatting}
\alias{reactable_formatting}
\title{Render reactable with certain formatting}
\usage{
reactable_formatting(
  dat,
  upper_case_title = TRUE,
  sortable = TRUE,
  sticky_cols = NULL,
  style = NULL,
  columns = NULL,
  minWidth = NULL,
  stock_type = "NSE",
  ...
)
}
\arguments{
\item{dat}{dataframe, data to render}

\item{upper_case_title}{boolean, if TRUE, will capitalise column names}

\item{sortable}{boolean, if TRUE, columns can be sorted}

\item{sticky_cols}{single string, name of column to make sticky}

\item{style}{name list of CSS params to apply to table}

\item{columns}{passed to reactable's column argument, see [reactable()]}

\item{minWidth}{integer, minimum width of each column}

\item{stock_type}{Stock Type NSE, NYSE or BSE}

\item{...}{additional arguments to pass to [reactable()]}
}
\value{
reactable table
}
\description{
This function renders reactable with custom formatting. The purpose of this
function is to standardize the formatting across the output tables.
}
\examples{
reactable_formatting(mtcars, sticky_cols = 'mpg')

}
