% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output_formating.R
\name{reactable_gen_theme}
\alias{reactable_gen_theme}
\title{Common theme for reactable's in RMA UI}
\usage{
reactable_gen_theme()
}
\value{
a theme option object
}
\description{
The purpose of this function is to have a customized default styling
for the output table
}
