% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_send_otp.R
\name{send_email}
\alias{send_email}
\title{Function to send mail and notification}
\usage{
send_email(user_email, username, password)
}
\arguments{
\item{user_email}{User mail id}

\item{username}{username}

\item{password}{password}
}
\value{
send notification based on email success or failure
}
\description{
Function to send mail and notification
}
