% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_send_otp.R
\name{send_otp_sms}
\alias{send_otp_sms}
\title{Function to send OTP on register mobile number}
\usage{
send_otp_sms(to_number, otp_code)
}
\arguments{
\item{to_number}{mobile number}

\item{otp_code}{OTP}
}
\value{
send the OTP to user's register mobile number
}
\description{
Function to send OTP on register mobile number
}
