% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output_formating.R
\name{sticky_style}
\alias{sticky_style}
\title{Utility to freeze the Period column
Make reactable column sticky}
\usage{
sticky_style(left = TRUE)
}
\arguments{
\item{left}{boolean}
}
\description{
Utility to freeze the Period column
Make reactable column sticky
}
