% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{ts_format_data}
\alias{ts_format_data}
\title{Function to convert data in TS format}
\usage{
ts_format_data(df, months_back = 0)
}
\arguments{
\item{df}{data frame}

\item{months_back}{months to look back}
}
\value{
return a time series data
}
\description{
Function to convert data in TS format
}
