% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validate_user.R
\name{validate_reset_pass}
\alias{validate_reset_pass}
\title{FUnction to reset the password}
\usage{
validate_reset_pass(
  session,
  user_email,
  new_pass,
  confirm_new_pass,
  users_data,
  on_success
)
}
\arguments{
\item{session}{current session}

\item{user_email}{user mail address}

\item{new_pass}{New password}

\item{confirm_new_pass}{Confirm New Password}

\item{users_data}{Users Data}

\item{on_success}{else part on success}
}
\value{
return notification if something is wrong
}
\description{
FUnction to reset the password
}
