% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validate_user.R
\name{validate_signup_user}
\alias{validate_signup_user}
\title{Function to validate user info while signing up}
\usage{
validate_signup_user(
  session,
  user_first_name,
  user_last_name,
  user_email,
  user_mob,
  users_data,
  on_success
)
}
\arguments{
\item{session}{session}

\item{user_first_name}{First name of user}

\item{user_last_name}{Last name of user}

\item{user_email}{Email ID of user}

\item{user_mob}{Mobile number of user}

\item{users_data}{Users data}

\item{on_success}{else part on success}
}
\value{
return notifications if something wrong
}
\description{
Function to validate user info while signing up
}
